\name{tidy}
\alias{tidy}
\title{Tidy a Model Object}
\description{
Generic function to convert model objects into tidy data frames.
}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{A model object.}
\item{...}{Additional arguments passed to methods.}
}
\value{
A data frame.
}
\seealso{
\code{\link{tidy.mrwin_fit}}
}
